+++ android/src/main/kotlin/com/superwall/superwallkit_flutter/BridgingCreator.kt
@@ -30,9 +30,8 @@ class BridgingCreator(
     object Constants {}
 
     companion object {
-        private var _shared: BridgingCreator? = null
-        val shared: BridgingCreator
-            get() = _shared ?: throw IllegalStateException("BridgingCreator not initialized")
+        private var _shared: MutableStateFlow<BridgingCreator?> = MutableStateFlow(null)
+        suspend fun shared() : BridgingCreator = _shared.filterNotNull().first()
 
         private var _flutterPluginBinding: MutableStateFlow<FlutterPlugin.FlutterPluginBinding?> =
             MutableStateFlow(null)
@@ -45,18 +44,14 @@ class BridgingCreator(
             // times (due to other SDK interference), we'll lose access to the
             // SuperwallKitFlutterPlugin current activity
             if (_flutterPluginBinding.value != null) {
-                SuperwallkitFlutterPlugin.reattachementCount.incrementAndGet()
                 println("WARNING: Attempting to set a flutter plugin binding again.")
                 return
             }
 
-            // Store for getter
-
-
             binding?.let {
                 synchronized(BridgingCreator::class.java) {
                     val bridge = BridgingCreator({ waitForPlugin() })
-                    _shared = bridge
+                    _shared.value = bridge
                     _flutterPluginBinding.value = binding
                     val communicator = Communicator(binding.binaryMessenger, "SWK_BridgingCreator")
                     communicator.setMethodCallHandler(bridge)
@@ -68,7 +63,7 @@ class BridgingCreator(
 
     fun tearDown() {
         print("Did tearDown BridgingCreator")
-        _shared = null
+        _shared.value = null
         _flutterPluginBinding.value = null
     }
 
@@ -123,6 +118,7 @@ class BridgingCreator(
             bridgeId,
             initializationArgs
         )
+
         bridgeInstance?.let { bridgeInstance ->
             instances[bridgeId] = bridgeInstance
             bridgeInstance.communicator().setMethodCallHandler(bridgeInstance)